requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-community
langchain-ollama
chromadb
anthropic
mcp
python-dotenv
pydantic
